pyproject.toml
[tool.poetry]
name = "whispipe"
version = "0.1.0"
description = "pipe voice control to stdout"
authors = ["7jane <7janeta@proton.me>"]
readme = "README.md"
packages = [{include = "whispipe"}]

[tool.poetry.dependencies]
python = "^3.11"
numba = "^0.60.0"
openai-whisper = "^20240930"
pydub = "^0.25.1"
torch = { file = "wheels/torch-2.4.1-rocm6.1-cp312-cp312-linux_x86_64.whl" }
torchvision = { file = "wheels/torchvision-0.19.1-rocm6.1-cp312-cp312-linux_x86_64.whl" }
torchaudio = { file = "wheels/torchaudio-2.4.1-rocm6.1-cp312-cp312-linux_x86_64.whl" }
pytorch-triton-rocm = { file = "wheels/pytorch_triton_rocm-3.0.0-cp312-cp312-linux_x86_64.whl" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
